uboot-sign-aspeed.bbclass: fix build failed if create bootloader fit image with SPL disable
If users disable to build SPL and enable to create bootloader fit image, it would build failed in uboot_fitimage_assemble function.
```
UBOOT_FITIMAGE_ENABLE = "1
```

Root Cause: SPL not found.
Solution: add to check SPL. If it exists, install it as signed SPL.

+++ meta-aspeed-sdk/classes/uboot-sign-aspeed.bbclass
@@ -405,7 +405,9 @@ EOF
 			${SPL_MKIMAGE_SIGN_ARGS}
 	fi
 
-	cp ${SPL_DIR}/${SPL_DTB_BINARY} ${SPL_DIR}/${SPL_DTB_SIGNED}
+	if [ -f "${SPL_DIR}/${SPL_NODTB_BINARY}" ]; then
+		cp ${SPL_DIR}/${SPL_DTB_BINARY} ${SPL_DIR}/${SPL_DTB_SIGNED}
+	fi
 }
 
 uboot_assemble_fitimage_helper() {
